OO-6514: change pos to Integer, patch XStream for 17

+++ src/main/java/org/olat/modules/coach/ui/curriculum/certificate/CurriculumTreeWithViewsRow.java
@@ -327,7 +327,7 @@ public class CurriculumTreeWithViewsRow implements CurriculumTreeWithView, Flexi
 		return element == null ? null : element.getEndDate();
 	}
 	
-	public Long getCurriculumElementPos() {
+	public Integer getCurriculumElementPos() {
 		return element == null ? null : element.getPos();
 	}
